refactor: merge default partial codecs into generic `CodecPartialDefault`
